Prelim fix to loginId problem while user is logged in and goes to /
