+++ app/main.jsx
@@ -10,6 +10,7 @@ import SignUp from './components/SignUp'
 import App from './components/App'
 import InlineBuddyEditIndex from './components/InlineBuddyEditIndex'
 import firebase from 'APP/fire'
+import redirectToFirstTrip from 'APP/src/redirectToFirstTrip'
 // ---- UNCOMMENT TO RESEED DATABASE ----------
 // import {trips, users} from '../seedData'
 // console.log('USERS', users, 'TRIPS', trips)
@@ -31,10 +32,15 @@ export default class Container extends React.Component {
   }
   componentWillMount() {
     this.unsubscribe = auth.onAuthStateChanged(user => {
-      // console.log('CONTAINER COMPONENT_WILL_MOUNT, USER: ', user)
+      console.log('CONTAINER COMPONENT_WILL_MOUNT, USER: ', user)
       this.setState({
         user: user
       })
+      if (!user) {
+        browserHistory.push('/login')
+      } else {
+        redirectToFirstTrip(user.uid)
+      }
     })
   }
   componentWillUnmount() {
@@ -42,10 +48,14 @@ export default class Container extends React.Component {
   }
   render() {
     // console.log('CONTAINER in MAIN, currentUser', firebase.auth().currentUser)
+    // Possible solution for no tripId bug below: If they were still logged in when navigated away
+    // but there is no tripId query string or route param,  then do what?
+    // Solution: make a utility fn that looks up a tripId
+      // for a previously logged in user and sets it bellow.
     return (
       <div className="container-fluid">
         {
-          this.state.user ? <App /> : <Login />
+          this.state.user ? null : <Login />
         }
       </div>
     )
